requirements.txt
requests
python-dotenv
slackweb
slackclient
oauth2